Refactor company admin view into modular components
